+++ src/__mocks__/server.js
@@ -1,6 +1,6 @@
 // src/mocks/server.js
 import { setupServer } from 'msw/node'
-import { handlers } from './handlers'
+import { authHandlers } from './auth-handlers'
 
 // This configures a request mocking server with the given request handlers.
-export const server = setupServer(...handlers)
+export const server = setupServer(...authHandlers)
